requirements.txt
click
numpy
pillow
tqdm
transformers
scikit-learn
torch
matplotlib
pandas
PySimpleGUI